src/inference-server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
paho-mqtt>=2.1.0
momentfm>=1.0.0
torch>=2.2.0
numpy>=1.26.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
